requirements.txt
flask~=3.0.2
flask-cors==4.0.0
flask-session==0.5.0
openai~=1.70.0
python-dotenv~=1.0.1
werkzeug~=3.1.3
pymupdf==1.23.3
python-docx==0.8.11
markupsafe==2.1.3
gunicorn==21.2.0
pdfplumber~=0.11.6
PyYAML~=6.0.1
langchain~=0.3.23